move *Server methods out of auth_with_roles.go (#10416)
